In progress: outside of map buidlings checking function.

+++ Assets/Scripts/Building.cs
@@ -28,7 +28,7 @@ public class Building
             if (Map.canBuildBuilding(this, y, x))
             {
                 //Building
-                id = (int)freeId;
+                id = (long)freeId;
                 uuid = cuuid;
             }else
             {
